Move datahub apps onto user model
