+++ users/models.py
@@ -1,3 +1,5 @@
+from .datahub import get_visible_apps
+
 from utils.models import APIModel
 
 
@@ -8,6 +10,8 @@ class Group(APIModel):
 
 
 class User(APIModel):
+    _apps = None
+
     @property
     def groups_display(self):
         group_names = [group["name"] for group in self.data.get("groups")]
@@ -27,3 +31,19 @@ class User(APIModel):
         if not self.is_active:
             return False
         return self.is_superuser or permission_name in self.permissions
+
+    def get_apps(self):
+        return get_visible_apps(self)
+
+    @property
+    def apps(self):
+        if self._apps is None:
+            self._apps = self.get_apps()
+        return self._apps
+
+    @property
+    def has_crm_permission(self):
+        for app in self.apps:
+            if app["permittedKey"] == "datahub-crm":
+                return True
+        return False
